+++ lib/src/emoji_picker_utils.dart
@@ -4,13 +4,13 @@ import 'dart:math';
 import 'package:emoji_picker_flutter/emoji_picker_flutter.dart';
 import 'package:emoji_picker_flutter/src/recent_emoji.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'emoji_lists.dart' as $emoji_list;
 
-/// EmojiPicker public method class
-/// Easy to use and expand other functions
-/// Reduce repeated encapsulation of the same function
+/// Helper class that provides extended usage
 
 class EmojiPickerUtils {
+  static Map<Category, Map<String, String>?>? _availableCategoryEmoji;
+  static final List<Emoji> _allAvailableEmojiEntities = [];
+
   /// Returns list of recently used emoji from cache
   static Future<List<RecentEmoji>> getRecentEmojis() async {
     final prefs = await SharedPreferences.getInstance();
@@ -22,24 +22,71 @@ class EmojiPickerUtils {
     return json.map<RecentEmoji>(RecentEmoji.fromJson).toList();
   }
 
-  /// Returns list of all the category emojis
-  static Map<Category, Map<String, String>> get getCategoryEmoji =>
-      _categoryEmoji;
+  /// Returns map of all the available category emojis
+  static Future<Map<Category, Map<String, String>?>?>
+      getAvailableCategoryEmoji() async {
+    if (_availableCategoryEmoji == null) {
+      // Filter recent category emojis
+      final categories =
+          Category.values.where((category) => category != Category.RECENT);
 
-  static final Map<Category, Map<String, String>> _categoryEmoji = Map.from({
-    Category.ACTIVITIES: $emoji_list.activities,
-    Category.ANIMALS: $emoji_list.animals,
-    Category.FLAGS: $emoji_list.flags,
-    Category.FOODS: $emoji_list.foods,
-    Category.OBJECTS: $emoji_list.objects,
-    Category.SMILEYS: $emoji_list.smileys,
-    Category.SYMBOLS: $emoji_list.symbols,
-    Category.TRAVEL: $emoji_list.travel
-  });
+      // Get all available emojis from cached
+      final futures = categories.map((category) async {
+        final emojis = await restoreFilteredEmojis(category.name);
+        return emojis;
+      });
+
+      final allAvailableEmojis = await Future.wait(futures);
+
+      _availableCategoryEmoji =
+          Map.fromIterables(categories, allAvailableEmojis);
+    }
+    return _availableCategoryEmoji;
+  }
+
+  /// Returns list of all the emoji entities
+  static Future<List<Emoji>> getAllAvailableEmojiEntities() async {
+    if (_allAvailableEmojiEntities.isEmpty) {
+      final availableCategoryEmoji = await getAvailableCategoryEmoji();
+      // Set all the emoji entities
+      availableCategoryEmoji?.forEach((_, emojis) {
+        final emojiEntities =
+            emojis?.entries.map((emoji) => Emoji(emoji.key, emoji.value));
+        _allAvailableEmojiEntities.addAll(emojiEntities ?? []);
+      });
+    }
+    return _allAvailableEmojiEntities;
+  }
+
+  /// Search for related emoticons based on keywords
+  static Future<List<Emoji>> searchEmoji(String keyword) async {
+    if (keyword.isEmpty) {
+      return [];
+    }
+
+    final allAvailableEmojiEntities = await getAllAvailableEmojiEntities();
+
+    return allAvailableEmojiEntities
+        .where((emoji) => emoji.name.contains(keyword))
+        .toList();
+  }
+
+  /// Restore locally cached emoji
+  static Future<Map<String, String>?> restoreFilteredEmojis(
+      String title) async {
+    final prefs = await SharedPreferences.getInstance();
+    var emojiJson = prefs.getString(title);
+    if (emojiJson == null) {
+      return null;
+    }
+    var emojis =
+        Map<String, String>.from(jsonDecode(emojiJson) as Map<String, dynamic>);
+    return emojis;
+  }
 
   /// Add an emoji to recently used list or increase its counter
   static Future<List<RecentEmoji>> addEmojiToRecentlyUsed(
-      {required Emoji emoji, int? recentsLimit}) async {
+      {required Emoji emoji, int recentsLimit = 1000}) async {
     final prefs = await SharedPreferences.getInstance();
     var recentEmoji = await getRecentEmojis();
     var recentEmojiIndex =
@@ -54,8 +101,7 @@ class EmojiPickerUtils {
     // Sort by counter desc
     recentEmoji.sort((a, b) => b.counter - a.counter);
     // Limit entries to recentsLimit
-    recentEmoji =
-        recentEmoji.sublist(0, min(recentsLimit!, recentEmoji.length));
+    recentEmoji = recentEmoji.sublist(0, min(recentsLimit, recentEmoji.length));
     // save locally
     prefs.setString('recent', jsonEncode(recentEmoji));